[GUI] Move SchemaArrayInput to a file
