+++ mc_rtc_rviz_panel/src/CMakeLists.txt
@@ -16,7 +16,7 @@ set(MOC_HEADERS
   NumberSliderWidget.h
   PlotWidget.h
   PlotTabWidget.h
-  Schema.h
+  SchemaArrayInput.h
   SchemaWidget.h)
 
 if(NOT ${DISABLE_ROS})
@@ -67,6 +67,7 @@ set(HEADER_FILES
   ClientWidget.h
   ConnectionConfiguration.h
   LabelWidget.h
+  Schema.h
   utils.h)
 
 if(NOT ${DISABLE_ROS})
@@ -92,6 +93,7 @@ set(SOURCE_FILES
   PlotWidget.cpp
   PlotTabWidget.cpp
   Schema.cpp
+  SchemaArrayInput.cpp
   SchemaWidget.cpp
   ${MOC_FILES}
   ${QRC_FILES})
